+++ src/main/scala/LagTracker.scala
@@ -48,7 +48,7 @@ final case class LagTracker(
   def lagMean: Option[Centis] = Option.when(steps > 0)(Centis(lagStats.mean))
 
   def compEstStdErr: Option[Float] =
-    Option.when(steps > 2)(math.sqrt(compEstSqErr).toFloat / (steps - 2))
+    Option.when(steps > 2)(math.sqrt(compEstSqErr.toDouble).toFloat / (steps - 2))
 
   def compAvg: Option[Centis] = totalComp / steps
